+++ examples/data-sources/catalog_entity_custom_data/resource.tf
@@ -1,0 +1,4 @@
+data "cortex_catalog_entity_custom_data" "products-region" {
+  tag = "products-service"
+  key = "region"
+}