+++ internal/provider/provider.go
@@ -117,6 +117,7 @@ func (p *CortexProvider) DataSources(ctx context.Context) []func() datasource.Da
 		NewDepartmentDataSource,
 		NewScorecardDataSource,
 		NewResourceDefinitionDataSource,
+		NewCatalogEntityCustomDataDataSource,
 	}
 }