+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog for the Cortex terraform provider.
 
 ## Unreleased
 
+* Add `cortex_catalog_entity_custom_data` data source for retrieving scoped custom data on catalog entities
 * Add `cortex_catalog_entity_custom_data` resource for managing scoped custom data on catalog entities
 
 ## 0.1.6